requirements.txt
alembic==1.13.3
aniso8601==9.0.1
async-timeout==4.0.3
bidict==0.23.1
click==8.1.7
commonmark==0.9.1
Deprecated==1.2.14
dnspython==2.2.1
eventlet==0.37.0
Flask==2.1.1
Flask-Limiter==3.8.0
Flask-Login==0.6.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
Mako==1.3.5
MarkupSafe==3.0.2
ordered-set==4.1.0
packaging==24.1
psycopg2-binary==2.9.3
Pygments==2.18.0
python-dotenv==0.21.0
python-engineio==4.10.1
python-socketio==5.4.0
redis==4.5.1
rich==12.6.0
simple-websocket==1.1.0
six==1.16.0
SQLAlchemy==1.4.32
typing_extensions==4.12.2
Werkzeug==2.0.3
wrapt==1.16.0
wsproto==1.2.0
Flask-Caching
openpyxl
reportlab
pdfkit
flask-marshmallow
flask_wtf
email_validator
sqlalchemy-utils
python-slugify